Fixed Image Edit Delete Issue

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
+        User currentUser= (User) session.getAttribute("loggeduser");
+        String error= "Only the owner of the image can edit the image";
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if (currentUser.equals(image.getUser())){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else {
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError", error);
+            return  "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -114,6 +121,7 @@ public class ImageController {
     //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
     public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
+
 
         Image image = imageService.getImage(imageId);
         String updatedImageData = convertUploadedFileToBase64(file);
@@ -140,11 +148,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
-    }
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session, Model model) {
+        Image image=imageService.getImage(imageId);
+        String error= "Only the owner of the image can delete the image";
+        User currentUser= (User) session.getAttribute("loggeduser");
+        if (currentUser.equals(image.getUser())){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else{
+            model.addAttribute("image",image);
+            model.addAttribute("deleteError", error);
+            return  "images/image";
+        }
 
+    }
 
     //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
